requirements_pip.txt
apispec==6.3.0
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-polymorphic==3.1.0
django-serializer==1.0.0
djangorestframework==3.14.0
ez-setup==0.9
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
